+++ phoenix/map/__init__.py
@@ -7,7 +7,7 @@ from pyramid.events import NewRequest
 from mako.template import Template
 from owslib.wms import WebMapService
 
-from twitcher.registry import service_registry_factory
+from phoenix.twitcherclient import twitcher_service_factory
 
 import logging
 logger = logging.getLogger(__name__)
@@ -21,20 +21,16 @@ map_script = Template(
 class Map(object):
     def __init__(self, request):
         self.request = request
-        self.session = self.request.session
 
     def view(self):
         dataset = self.request.params.get('dataset')
         wms_url = self.request.params.get('wms_url')
         if dataset:
-            use_proxy = False
             url = self.request.route_url(
-                'owsproxy',
-                service_name='wms',
+                'wms',
                 _query=[('DATASET', dataset)])
             caps_url = self.request.route_url(
-                'owsproxy',
-                service_name='wms',
+                'wms',
                 _query=[('DATASET', dataset),
                         ('service', 'WMS'), ('request', 'GetCapabilities'), ('version', '1.1.1')])
             try:
@@ -47,7 +43,6 @@ class Map(object):
                 logger.exception("wms connect failed")
                 raise Exception("could not connect to wms url %s", caps_url)
         elif wms_url:
-            use_proxy = True
             try:
                 wms = WebMapService(wms_url)
                 map_name = wms_url.split('/')[-1]
@@ -57,8 +52,7 @@ class Map(object):
         else:
             wms = None
             map_name = None
-            use_proxy = False
-        return dict(map_script=map_script.render(wms=wms, dataset=dataset, use_proxy=use_proxy),
+        return dict(map_script=map_script.render(wms=wms, dataset=dataset),
                     map_name=map_name)
 
 
@@ -69,28 +63,9 @@ def includeme(config):
         return asbool(settings.get('phoenix.map', 'false'))
     config.add_request_method(map_activated, reify=True)
 
-    def wms_url(request):
-        return settings.get('wms.url')
-    config.add_request_method(wms_url, reify=True)
-
     if asbool(settings.get('phoenix.map', 'false')):
-        # add wms service
-        def add_wms(event):
-            request = event.request
-            if request.map_activated and not 'wms' in settings:
-                logger.debug('register wms service')
-                try:
-                    service_name = 'wms'
-                    registry = service_registry_factory(request.registry)
-                    logger.debug("register: name=%s, url=%s", service_name, settings['wms.url'])
-                    registry.register_service(name=service_name, url=settings['wms.url'],
-                                              public=True, service_type='wms',
-                                              overwrite=True)
-                    settings['wms'] = WebMapService(url=settings['wms.url'])
-                except:
-                    logger.exception('Could not connect wms %s', settings['wms.url'])
-            event.request.wms = settings.get('wms')
-        config.add_subscriber(add_wms, NewRequest)
+        # wms url
+        config.add_route('wms', settings['wms.url'])
 
         # map view
         config.add_route('map', '/map')