+++ phoenix/monitor/panels/inputs.py
@@ -3,7 +3,7 @@ from pyramid_layout.panel import panel_config
 from phoenix.wps import check_status
 
 import logging
-logger = logging.getLogger(__name__)
+LOGGER = logging.getLogger(__name__)
 
 
 def collect_inputs(status_location=None, response=None):
@@ -31,10 +31,12 @@ class Inputs(object):
     @panel_config(name='monitor_inputs', renderer='../templates/monitor/panels/media.pt')
     def panel(self):
         job_id = self.request.matchdict.get('job_id')
+        wps_output_url = self.request.registry.settings.get('wps.output.url')
 
         items = []
         for inp in process_inputs(self.request, job_id):
             dataset = None
+            proxy_reference = inp.reference
             # TODO: use config for nwms dynamic services
             if self.request.map_activated and inp.mimeType and 'netcdf' in inp.mimeType and inp.reference:
                 if 'cache' in inp.reference:
@@ -49,6 +51,11 @@ class Inputs(object):
                     dataset = "archive-cordex" + inp.reference.split('CORDEX/data')[1]
                 elif 'OBS4MIPS/data' in inp.reference:
                     dataset = "archive-obs4mips" + inp.reference.split('OBS4MIPS/data')[1]
+            if inp.reference and wps_output_url and inp.reference.startswith(wps_output_url):
+                proxy_reference = self.request.route_url(
+                    'download_wpsoutputs',
+                    subpath=inp.reference.split(wps_output_url)[1])
+                LOGGER.debug("proxy reference: %s", proxy_reference)
             if inp.mimeType:
                 category = 'ComplexType'
             else:
@@ -60,6 +67,7 @@ class Inputs(object):
                               mime_type=inp.mimeType,
                               data=inp.data,
                               reference=inp.reference,
+                              proxy_reference=proxy_reference,
                               dataset=dataset,
                               category=category))