+++ phoenix/monitor/panels/outputs.py
@@ -34,17 +34,27 @@ class Outputs(object):
     @panel_config(name='monitor_outputs', renderer='../templates/monitor/panels/media.pt')
     def panel(self):
         job_id = self.request.matchdict.get('job_id')
+        wps_output_url = self.request.registry.settings.get('wps.output.url')
 
         items = []
         for output in process_outputs(self.request, job_id).values():
             dataset = None
-            if self.request.map_activated and output.mimeType and 'netcdf' in output.mimeType:
-                if output.reference and 'wpsoutputs' in output.reference:
+            proxy_reference = output.reference
+            logger.debug("output reference: %s", output.reference)
+            if output.reference and 'wpsoutputs' in output.reference:
+                if self.request.map_activated and output.mimeType and 'netcdf' in output.mimeType:
                     dataset = "outputs" + output.reference.split('wpsoutputs')[1]
+                if wps_output_url and output.reference.startswith(wps_output_url):
+                    proxy_reference = self.request.route_url(
+                        'download_wpsoutputs',
+                        subpath=output.reference.split(wps_output_url)[1])
+                    logger.debug("proxy reference: %s", proxy_reference)
             if output.mimeType:
                 category = 'ComplexType'
             else:
                 category = 'LiteralType'
+
+            logger.debug("proxy_reference: %s", proxy_reference)
 
             items.append(dict(title=output.title,
                               abstract=output.abstract,
@@ -52,6 +62,7 @@ class Outputs(object):
                               mime_type=output.mimeType,
                               data=output.data,
                               reference=output.reference,
+                              proxy_reference=proxy_reference,
                               dataset=dataset,
                               category=category))
         items = sorted(items, key=lambda item: item['identifier'], reverse=1)