+++ phoenix/monitor/views/actions.py
@@ -1,13 +1,16 @@
+import requests
+
 from pyramid.view import view_config, view_defaults
 from pyramid.httpexceptions import HTTPFound
 from pyramid.security import authenticated_userid
+from pyramid.response import Response
+from pyramid.response import FileIter
 
 from phoenix.utils import ActionButton
 from phoenix.utils import format_tags
 
 import logging
-
-logger = logging.getLogger(__name__)
+LOGGER = logging.getLogger(__name__)
 
 
 @view_defaults(permission='submit')
@@ -159,6 +162,51 @@ def monitor_buttons(context, request):
     return buttons
 
 
+def download_wpsoutputs(request):
+    # TODO: streaming html files does not work ...
+    path = '/'.join(request.matchdict['subpath'])
+    url = request.registry.settings.get('wps.output.url')
+    url += '/' + path
+    LOGGER.debug("delegate to wpsoutputs: %s", url)
+    # forward request to target (without Host Header)
+    # h = dict(request.headers)
+    # h.pop("Host", h)
+    response = requests.get(url, stream=True, verify=False)
+    response.raise_for_status()
+
+    def remove_header(key, headers):
+        # the default header key should be the standard capitilized version e.g 'Content-Length'
+        #TODO: move code to twitcher owsproxy
+        try:
+            del headers[key]
+        except KeyError:
+            try:
+                del headers[key.lower()]
+            except KeyError:
+                try:
+                    del headers[key.upper()]
+                except KeyError:
+                    pass
+
+    # clean up headers
+    headers = dict(response.headers)
+    keys = [k.lower() for k in headers.keys()]
+    if 'content-length' in keys:
+        remove_header('Content-Length', headers)
+    if 'transfer-encoding' in keys:
+        remove_header('Transfer-Encoding', headers)
+    if 'content-encoding' in keys:
+        remove_header('Content-Encoding', headers)
+    if 'connection' in keys:
+        remove_header('Connection', headers)
+    if 'keep-alive' in keys:
+        remove_header('Keep-Alive', headers)
+
+    proxy_response = Response(app_iter=FileIter(response.raw), status=response.status_code)
+    proxy_response.headers.update(headers)
+    return proxy_response
+
+
 def includeme(config):
     """ Pyramid includeme hook.
 
@@ -174,3 +222,6 @@ def includeme(config):
     config.add_route('make_private', 'make_private')
     config.add_route('set_favorite', 'set_favorite')
     config.add_route('unset_favorite', 'unset_favorite')
+    # download internal wps outputs
+    config.add_route('download_wpsoutputs', '/download/wpsoutputs*subpath')
+    # config.add_view(download_wpsoutputs, route_name='download_wpsoutputs')