+++ phoenix/people/views/actions.py
@@ -1,9 +1,9 @@
 from pyramid.view import view_config
-
 from pyramid.httpexceptions import HTTPFound
 from pyramid.security import authenticated_userid
 
-from phoenix.security import generate_access_token
+from phoenix.twitcherclient import generate_access_token
+from phoenix.esgfslcs import ESGFSLCSClient
 
 import logging
 logger = logging.getLogger(__name__)
@@ -13,6 +13,7 @@ class Actions(object):
     def __init__(self, request):
         self.request = request
         self.session = request.session
+        settings = request.registry.settings
         self.collection = self.request.db.users
         self.userid = self.request.matchdict.get('userid', authenticated_userid(self.request))
 
@@ -27,8 +28,34 @@ class Actions(object):
 
     @view_config(route_name='generate_twitcher_token', permission='submit')
     def generate_twitcher_token(self):
-        generate_access_token(self.request.registry, self.userid)
+        generate_access_token(self.request.registry, userid=self.userid)
         return HTTPFound(location=self.request.route_path('profile', userid=self.userid, tab='twitcher'))
+
+    @view_config(route_name='generate_esgf_slcs_token', permission='submit')
+    def generate_esgf_slcs_token(self):
+        """
+        Update ESGF slcs token.
+        """
+        client = ESGFSLCSClient(self.request)
+        if client.get_token():
+            client.refresh_token()
+            return HTTPFound(location=self.request.route_path('profile', userid=self.userid, tab='esgf_slcs'))
+        else:
+            auth_url = client.authorize()
+            return HTTPFound(location=auth_url)
+
+    @view_config(route_name='esgf_oauth_callback', permission='submit')
+    def esgf_oauth_callback(self):
+        """
+        Convert an authorisation grant into an access token.
+        """
+        client = ESGFSLCSClient(self.request)
+        if client.callback():
+            # Redirect to the token view
+            return HTTPFound(location=self.request.route_path('profile', userid=self.userid, tab='esgf_slcs'))
+        else:
+            # If we have not yet entered the OAuth flow, redirect to the start
+            return HTTPFound(location=self.request.route_path('generate_esgf_slcs_token'))
 
     @view_config(route_name='delete_user', permission='admin')
     def delete_user(self):
@@ -45,4 +72,6 @@ def includeme(config):
     """
     config.add_route('forget_esgf_certs', 'people/forget_esgf_certs')
     config.add_route('generate_twitcher_token', 'people/gentoken')
+    config.add_route('generate_esgf_slcs_token', 'people/generate_esgf_token')
+    config.add_route('esgf_oauth_callback', 'account/oauth/esgf/callback')
     config.add_route('delete_user', 'people/delete/{userid}')