+++ phoenix/security.py
@@ -4,7 +4,7 @@ see pyramid security:
 * http://docs.pylonsproject.org/projects/pyramid/en/latest/tutorials/wiki2/authentication.html
 """
 
-from datetime import datetime
+from collections import OrderedDict
 
 from pyramid.authentication import AuthTktAuthenticationPolicy
 from pyramid.authorization import ACLAuthorizationPolicy
@@ -17,15 +17,9 @@ from pyramid.security import (
 
 from authomatic import Authomatic, provider_id
 from authomatic.providers import oauth2, openid
-from phoenix.providers import oauth2 as myoauth2
 from phoenix.providers import esgfopenid
 
-
-from twitcher.tokens import tokengenerator_factory
-from twitcher.tokens import tokenstore_factory
-from twitcher.registry import service_registry_factory
-
-from phoenix.db import mongodb
+from phoenix.twitcherclient import is_public
 
 import logging
 logger = logging.getLogger(__name__)
@@ -34,36 +28,33 @@ Admin = 'group.admin'
 User = 'group.user'
 Guest = 'group.guest'
 
+AUTH_PROTOCOLS = OrderedDict([
+    ('phoenix', 'Phoenix'),
+    ('esgf', 'ESGF OpenID'),
+    ('openid', 'OpenID'),
+    ('oauth2', 'OAuth 2.0'),
+    ('ldap', 'LDAP')])
+
 
 def has_execute_permission(request, service_name):
-    service_registry = service_registry_factory(request.registry)
-    return service_registry.is_public(service_name) or request.has_permission('submit')
+    return is_public(request.registry, service_name) or request.has_permission('submit')
 
 
-def generate_access_token(registry, userid=None):
-    db = mongodb(registry)
-
-    tokengenerator = tokengenerator_factory(registry)
-    access_token = tokengenerator.create_access_token(valid_in_hours=8, user_environ={})
-    tokenstore = tokenstore_factory(registry)
-    tokenstore.save_token(access_token)
-    token = access_token['token']
-    expires = datetime.utcfromtimestamp(
-        int(access_token['expires_at'])).strftime(format="%Y-%m-%d %H:%M:%S UTC")
-    if userid:
-        db.users.update_one({'identifier': userid},
-                            {'$set': {'twitcher_token': token, 'twitcher_token_expires': expires}})
-
-
-def auth_protocols(request):
+def allowed_auth_protocols(request):
     # TODO: refactor auth settings handling
-    settings = request.db.settings.find_one()
-    protocols = ['phoenix', 'esgf', 'openid', 'ldap', 'oauth2']
-    if settings:
-        if 'auth' in settings:
-            if 'protocol' in settings['auth']:
-                protocols = settings['auth']['protocol']
+    settings = request.db.settings.find_one() or {}
+    protocols = ['phoenix', 'esgf', 'oauth2']
+    if 'auth_protocol' in settings:
+        protocols.extend(settings['auth_protocol'])
     return protocols
+
+
+def default_auth_protocol(request):
+    allowed_protocols = allowed_auth_protocols(request)
+    # use reverse order to get defaul protocol
+    for protocol in AUTH_PROTOCOLS.keys()[::-1]:
+        if protocol in allowed_protocols:
+            return protocol
 
 
 def passwd_check(request, passphrase):
@@ -178,8 +169,8 @@ def authomatic_config(request):
     OAUTH2 = {
         'github': {
             'class_': oauth2.GitHub,
-            'consumer_key': request.github_oauth[0],
-            'consumer_secret': request.github_oauth[1],
+            'consumer_key': request.registry.settings.get('github.client.id'),
+            'consumer_secret': request.registry.settings.get('github.client.secret'),
             'access_headers': {'User-Agent': 'Phoenix'},
             'id': provider_id(),
             'scope': oauth2.GitHub.user_info_scope,
@@ -187,15 +178,6 @@ def authomatic_config(request):
                 'Get your events': ('GET', 'https://api.github.com/users/{user.username}/events'),
                 'Get your watched repos': ('GET', 'https://api.github.com/user/subscriptions'),
             },
-        },
-        'ceda': {
-            'class_': myoauth2.Ceda,
-            'consumer_key': request.registry.settings.get('ceda.consumer.key'),
-            'consumer_secret': request.registry.settings.get('ceda.consumer.secret'),
-            'id': provider_id(),
-            'scope': myoauth2.Ceda.user_info_scope,
-            #'state': 'ceda',
-            'redirect_uri': request.registry.settings.get('ceda.consumer.redirect.uri'),
         },
     }