+++ phoenix/settings/views/overview.py
@@ -18,10 +18,16 @@ class Overview(MyView):
 
         buttons.append(dict(url=self.request.route_path('services'), icon_class="fa fa-server fa-2x", title="Services"))
         if self.request.solr_activated:
-            buttons.append(dict(url=self.request.route_path('settings_solr', tab='index'), icon_class="fa fa-search fa-2x", title="Solr"))
-        buttons.append(dict(url=self.request.route_path('settings_auth'), icon_class="fa fa-lock fa-2x", title="Auth"))
-        buttons.append(dict(url=self.request.route_path('settings_ldap'), icon_class="fa fa-sitemap fa-2x", title="LDAP"))
-        buttons.append(dict(url=self.request.route_path('settings_github'), icon_class="fa fa-github fa-2x", title="GitHub"))
+            buttons.append(dict(url=self.request.route_path('settings_solr', tab='index'),
+                                icon_class="fa fa-search fa-2x", title="Solr"))
+        buttons.append(dict(url=self.request.route_path('settings_auth'),
+                            icon_class="fa fa-lock fa-2x", title="Auth"))
+        buttons.append(dict(url=self.request.route_path('settings_ldap'),
+                            icon_class="fa fa-sitemap fa-2x", title="LDAP"))
+        buttons.append(dict(url=self.request.route_path('settings_github'),
+                            icon_class="fa fa-github fa-2x", title="GitHub"))
+        buttons.append(dict(url=self.request.route_path('settings_esgf'),
+                            icon_class="fa fa-key fa-2x", title="ESGF SLCS"))
         buttongroups.append(dict(title='Settings', buttons=buttons))
 
         return dict(buttongroups=buttongroups)