+++ phoenix/wizard/views/start.py
@@ -11,6 +11,8 @@ from phoenix.wps import check_status
 from phoenix.utils import time_ago_in_words
 from phoenix.wizard.views import Wizard
 
+from twitcher.utils import parse_service_name
+
 import logging
 logger = logging.getLogger(__name__)
 
@@ -37,10 +39,11 @@ def job_to_state(request, job_id):
 
             # TODO: get the correct worker wps url
             # TODO: avoid getcaps
-            wps = WebProcessingService(url=workflow['worker']['url'], verify=False, skip_caps=False)
+            wps = WebProcessingService(url=workflow['worker']['url'].split('?')[0], verify=False, skip_caps=False)
             process = wps.describeprocess(workflow['worker']['identifier'])
 
-            state['wizard_wps'] = {'identifier': request.catalog.get_service_name_by_url(wps.url)}
+            service = request.catalog.get_service_by_url(wps.url)
+            state['wizard_wps'] = {'identifier': service['name']}
             state['wizard_process'] = {'identifier': workflow['worker']['identifier']}
             inputs = {}
             for inp in workflow['worker']['inputs']: